requirements.txt
selenium==4.31.0
pytest==8.3.5
translate==3.6.1
python-dotenv==1.1.0
requests==2.32.3
Pillow==11.2.1
